# docs/openapi.yaml
openapi: 3.0.3
info:
  title: EventHub API
  version: 0.2.0
  description: |
    MVP OpenAPI spec for EventHub.
    Focused on core auth, events, registrations, and admin operations.
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /events:
    get:
      tags: [Events]
      summary: List events
      operationId: listEvents
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/City"
        - $ref: "#/components/parameters/SearchQuery"
        - $ref: "#/components/parameters/From"
        - $ref: "#/components/parameters/To"
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/IncludeTotal"
        - $ref: "#/components/parameters/IfNoneMatch"
      responses:
        "200":
          description: Events page
          headers:
            ETag:
              schema:
                type: string
              description: Entity tag for conditional requests.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventListResponse"
              example:
                limit: 20
                count: 2
                items:
                  - id: 3d8e4d13-bad3-4fb7-9022-8b5fa77111d2
                    title: Go Backend Deep Dive
                    description: Build and ship a production Go API
                    city: Lagos
                    startAt: 2026-03-01T09:00:00Z
                    capacity: 120
                    createdAt: 2026-02-01T12:00:00Z
                    updatedAt: 2026-02-01T12:00:00Z
                  - id: 7f027825-9162-4d6e-ae66-6d6be2178698
                    title: Intro to Distributed Systems
                    description: Practical patterns for reliability
                    city: Abuja
                    startAt: 2026-03-03T10:30:00Z
                    capacity: 90
                    createdAt: 2026-02-02T12:00:00Z
                    updatedAt: 2026-02-02T12:00:00Z
                hasMore: true
                nextCursor: eyJzdGFydEF0IjoiMjAyNi0wMy0wM1QxMDozMDowMFoiLCJpZCI6IjdmMDI3ODI1LTkxNjItNGQ2ZS1hZTY2LTZkNmJlMjE3ODY5OCJ9
                total: null
        "400":
          $ref: "#/components/responses/Error"
        "304":
          description: Not Modified (matched `If-None-Match`)
        "500":
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    refreshTokenCookie:
      type: apiKey
      in: cookie
      name: refresh_token
  parameters:
...
    Limit:
      in: query
      name: limit
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    Cursor:
      in: query
      name: cursor
      required: false
      description: Base64url keyset cursor.
      schema:
        type: string
    IncludeTotal:
      in: query
      name: includeTotal
      required: false
      schema:
        type: boolean
        default: false
    IfNoneMatch:
      in: header
      name: If-None-Match
      required: false
      description: Revalidate a cached representation using an ETag value.
      schema:
        type: string
    City:
      in: query
      name: city
      required: false
      schema:
        type: string
    SearchQuery:
      in: query
      name: q
      required: false
      description: Full-text search across event `title`, `description`, and `city`.
      schema:
        type: string
    From:
      in: query
      name: from
      required: false
      description: RFC3339 timestamp lower bound for `startAt`.
      schema:
        type: string
        format: date-time
    To:
      in: query
      name: to
      required: false
      description: RFC3339 timestamp upper bound for `startAt`.
      schema:
        type: string
        format: date-time
...
  responses:
    Error:
      description: Error response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            generic:
              value:
                error:
                  code: invalid_request
                  message: Invalid request body
                  requestId: 4bca7777d14b4b1a
            validation:
              value:
                error:
                  code: invalid_request
                  message: Invalid request body
                  requestId: 4bca7777d14b4b1a
                  details:
                    fields:
                      - field: title
                        rule: min
                        param: "3"
                        message: must be at least 3
                      - field: startAt
                        rule: required
                        message: is required
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            requestId:
              type: string
              nullable: true
            details:
              nullable: true
              oneOf:
                - type: string
                - type: object
                  additionalProperties: true
                - type: array
                  items:
                    type: object
                    additionalProperties: true
...
    Event:
      type: object
      required: [id, title, startAt, capacity, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        city:
          type: string
        startAt:
          type: string
          format: date-time
        capacity:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    EventListResponse:
      type: object
      required: [limit, count, items, hasMore, nextCursor, total]
      properties:
        limit:
          type: integer
        count:
          type: integer
        items:
          type: array
          items:
            $ref: "#/components/schemas/Event"
        hasMore:
          type: boolean
        nextCursor:
          type: string
          nullable: true
        total:
          type: integer
          nullable: true
...